+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/SessionSuggestionProvider.kt
@@ -9,8 +9,9 @@ import kotlinx.coroutines.Deferred
 import mozilla.components.browser.icons.BrowserIcons
 import mozilla.components.browser.icons.Icon
 import mozilla.components.browser.icons.IconRequest
-import mozilla.components.browser.session.Session
-import mozilla.components.browser.session.SessionManager
+import mozilla.components.browser.state.state.BrowserState
+import mozilla.components.browser.state.state.TabSessionState
+import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.awesomebar.AwesomeBar
 import mozilla.components.feature.awesomebar.R
 import mozilla.components.feature.tabs.TabsUseCases
@@ -22,7 +23,7 @@ import java.util.UUID
  */
 class SessionSuggestionProvider(
     private val resources: Resources,
-    private val sessionManager: SessionManager,
+    private val store: BrowserStore,
     private val selectTabUseCase: TabsUseCases.SelectTabUseCase,
     private val icons: BrowserIcons? = null,
     private val excludeSelectedSession: Boolean = false
@@ -34,20 +35,26 @@ class SessionSuggestionProvider(
             return emptyList()
         }
 
-        val suggestions = mutableListOf<AwesomeBar.Suggestion>()
-        val iconRequests: List<Deferred<Icon>?> = sessionManager.sessions.map { icons?.loadIcon(IconRequest(it.url)) }
+        val state = store.state
+        val tabs = state.tabs
 
-        sessionManager.sessions.zip(iconRequests) { result, icon ->
-            if (result.contains(text) && !result.private && shouldIncludeSelectedSession(result)
+        val suggestions = mutableListOf<AwesomeBar.Suggestion>()
+        val iconRequests: List<Deferred<Icon>?> = tabs.map { icons?.loadIcon(IconRequest(it.content.url)) }
+
+        tabs.zip(iconRequests) { result, icon ->
+            if (
+                result.contains(text) &&
+                !result.content.private &&
+                shouldIncludeSelectedTab(state, result)
             ) {
                 suggestions.add(
                         AwesomeBar.Suggestion(
                             provider = this,
                             id = result.id,
-                            title = result.title,
+                            title = result.content.title,
                             description = resources.getString(R.string.switch_to_tab_description),
                             icon = icon?.await()?.bitmap,
-                            onSuggestionClicked = { selectTabUseCase(result) }
+                            onSuggestionClicked = { selectTabUseCase(result.id) }
                         )
                 )
             }
@@ -55,16 +62,14 @@ class SessionSuggestionProvider(
         return suggestions
     }
 
-    private fun Session.contains(text: String) =
-            (url.contains(text, ignoreCase = true) || title.contains(text, ignoreCase = true))
+    private fun TabSessionState.contains(text: String) =
+            (content.url.contains(text, ignoreCase = true) || content.title.contains(text, ignoreCase = true))
 
-    private fun shouldIncludeSelectedSession(session: Session): Boolean {
+    private fun shouldIncludeSelectedTab(state: BrowserState, tab: TabSessionState): Boolean {
         return if (excludeSelectedSession) {
-            !isSelectedSession(session)
+            tab.id != state.selectedTabId
         } else {
             true
         }
     }
-
-    private fun isSelectedSession(session: Session) = sessionManager.selectedSession?.equals(session) ?: false
 }